add database reading

+++ app/src/main/java/net/gas/contactbook/ui/database/entities/Posts.kt
@@ -5,7 +5,7 @@ import androidx.room.PrimaryKey
 
 @Entity(tableName = "posts")
 data class Posts (
-    @PrimaryKey val id: Int,
+    @PrimaryKey val id: Short,
     val name: String?,
     val rangir: Int?
 )
